% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{boot_summary}
\alias{boot_summary}
\title{Summary of bootstrapped statistics}
\usage{
boot_summary(x, ci_method = c("percentile", "bca"))
}
\arguments{
\item{x}{a numeric vector.}

\item{ci_method}{methods for calculation of confidence intervals:
\code{percentile} (default) or \code{bca} (bias corrected and accelerated).}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{boot_mean} and \code{boot_sd}: mean and standard deviation across
the bootstraps
\item \code{boot_median}: median bootstrapped value
\item \code{boot_lower_iqr} and \code{boot_upper_iqr}: lower and upper bound of the
interquartile range
\item \code{boot_lower_ci} and \code{boot_upper_ci}: lower and upper bound of the
95\% confidence interval
}
}
\description{
A default function used by the package to compute summary metrics of a
bootstrapped statistic.
}
\details{
Silently ignores any missing values.
}
