% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_repeated}
\alias{filter_repeated}
\title{Filter a data frame with repeated variable levels.}
\usage{
filter_repeated(data, variable, levels)
}
\arguments{
\item{data}{a data frame.}

\item{variable}{character specifying the variable name.}

\item{levels}{a character vector with levels of the variable which will
be selected for the output.}
}
\value{
a data frame.
}
\description{
Filters the input data frame by levels of a variable specified by the user
in a character vector called \code{levels}. If there are duplicated elements
in such character vector, they will be accounted for in the output
data frame: those levels will be duplicated es well. Note that this is
the difference between \code{filter_repeated()} and \code{dplyr::filter()}
}
\details{
Designed for internal use.
}
