% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{bmap}
\alias{bmap}
\alias{bmap.default}
\title{Bootstrapped descriptive statistic}
\usage{
bmap(x, B, FUN, ...)

\method{bmap}{default}(
  x,
  B,
  FUN,
  ci_method = c("percentile", "bca"),
  summary_function = NULL,
  ...
)
}
\arguments{
\item{x}{an object, currently atomic vectors, matrices and data frames
are implemented.}

\item{B}{number of bootstrapped samples.}

\item{FUN}{a user-provided function. It should return a numeric vector.}

\item{...}{extra arguments provided to the function \code{FUN}.}

\item{ci_method}{methods for calculation of confidence intervals:
\code{percentile} (default) or \code{bca} (bias corrected and accelerated).
Used by the \code{\link{boot_summary}} function.}

\item{summary_function}{a function used for calculation of the bootstrapped
metrics, can be virtually any user-provided function which takes
a numeric vector. If NULL, \code{\link{boot_summary}} is used.}
}
\value{
A list of two elements \code{dataset} and \code{bootstrap} which store the evaluation
result in the entire dataset and bootstrapped metrics, respectively.
}
\description{
Computes an estimate of the descriptive statistic (or statistics) returned
by a user-provided function in the genuine data set and in bootstrapped
samples of the data set.
}
\details{
The \code{bmap()} function is a S3 generic.
It will run in parallel when a parallel backend is registered
via \code{future::plan()}.
}
